850C eeprom stuff is mergemotized, but not yet tested
